move shopping db context to data namespace
