+++ src/Shopping/Server/Controllers/ProductCategoryController.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
+using Shopping.Server.Data;
 using Shopping.Shared.Data;
 
 namespace Shopping.Server.Controllers
