+++ src/Shopping/Server/Controllers/ShoppingListItemsController.cs
@@ -7,7 +7,7 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using Shopping.Server;
+using Shopping.Server.Data;
 using Shopping.Shared.Data;
 
 namespace Shopping.Server.Controllers
