+++ src/Shopping/Server/Data/ShoppingDbContext.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Shopping.Server
+namespace Shopping.Server.Data
 {
     public class ShoppingDbContext : DbContext
     {
