+++ src/Shopping/Server/Startup.cs
@@ -1,11 +1,9 @@
 using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.HttpsPolicy;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.ResponseCompression;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using System.Linq;
+using Shopping.Server.Data;
 using Microsoft.EntityFrameworkCore;
 
 namespace Shopping.Server
