+++ src/Shopping/Server/appsettings.Development.json
@@ -7,6 +7,7 @@
     }
   },
   "ConnectionStrings": {
-    "Shopping_Azure": "AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw=="
+    "Shopping_Azure": "AccountEndpoint=https://localhost:8081/;AccountKey=C2y6yDjf5/R+ob0N8A7Cgv30VRDJIWEHLM+4QDU5DE2nQ9nDuVTqobD4b8mGGyPMbIZnqyMsEcaGQy67XIw/Jw==",
+    "IdentityMySQL": "Server=127.0.0.1; Database=localdb; Uid=root; Password=Test123!#"
   }
 }
